backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.0
drf-extra-fields==3.7.0
django-filter==23.2
Pillow==10.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
